requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-mail
reportlab
python-dotenv